Refactor with Base class
